refactor: light refactor

+++ models/password_check.go
@@ -1,5 +1,5 @@
 package models
 
 type PasswordCheck struct {
-	Strength float64 `json:"strength"`
+	Strength int `json:"strength"`
 }
